requirements.txt
markdown>=3.5
pyyaml>=6.0
jinja2>=3.1
watchdog>=4.0
